repro build - build_cmd.rs rename
